+++ src/main/kotlin/com/github/pirocraft/challengingtimer/example/services/MyApplicationService.kt
@@ -1,0 +1,10 @@
+package com.github.pirocraft.challengingtimer.example.services
+
+import com.github.pirocraft.challengingtimer.example.MyBundle
+
+class MyApplicationService {
+
+    init {
+        println(MyBundle.message("applicationService"))
+    }
+}
