--- src/main/kotlin/com/github/pirocraft/challengingtimer/listeners/MyProjectManagerListener.kt
@@ -1,12 +1,0 @@
-package com.github.pirocraft.challengingtimer.listeners
-
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.ProjectManagerListener
-import com.github.pirocraft.challengingtimer.services.MyProjectService
-
-internal class MyProjectManagerListener : ProjectManagerListener {
-
-    override fun projectOpened(project: Project) {
-        project.getService(MyProjectService::class.java)
-    }
-}
