--- src/main/kotlin/com/github/pirocraft/challengingtimer/services/MyApplicationService.kt
@@ -1,10 +1,0 @@
-package com.github.pirocraft.challengingtimer.services
-
-import com.github.pirocraft.challengingtimer.MyBundle
-
-class MyApplicationService {
-
-    init {
-        println(MyBundle.message("applicationService"))
-    }
-}
